+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * **Option to randomize grand arena defenders**
 * **Add `grand_generate_solutions` and `grand_keep_solutions` options**
 * Add required parameter to `friendsSendDailyGift`
+* Update `User-Agent` and resources
 
 # `2.0`